Moved from I2C terminal simulator to the Wireless Telemetry interface
The I2C terminal interface did not work fast enough as the key presses were not accepted fast enough by the ECU. The wireless telemetry interface is based on serial communication and transmits all relevant data directly.

The main platform has been moved to the Teensy-LC to take advantage of the multiple hardware serial interfaces and the half duplex inverted capabilities in hardware.